summarizer-fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
fastapi==0.115.12
fastavro==1.10.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
jieba3k==0.35.1
jiter==0.9.0
joblib==1.4.2
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
openai==1.76.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
playwright==1.51.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
selenium==4.31.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
